Add recommendations for viscosity and velocity for K17 kernel and move the check to kernel

See merge request irmb/VirtualFluids!443

+++ tests/unit-tests/gpu/GridGenerator/grid/GridImpTest.cpp
@@ -1063,7 +1063,7 @@ void logForFindCommunicationIndexData(real coordinate, real coordinateLimit,
                                       CommunicationDirections::CommunicationDirection direction, real delta)
 {
     VF_LOG_INFO("coordinate = {},\tcoordinateLimit = {},\tdirection = {} (isPositive: {}),\tdelta = {}", coordinate, coordinateLimit,
-                direction, CommunicationDirections::isPositive(direction), delta);
+                static_cast<int>(direction), CommunicationDirections::isPositive(direction), delta);
 }
 
 TEST(GridImpTest, findCommunicationIndex_MX)
